+++ packages/psychart/example/Mollier Diagram/src/index.js
@@ -1,16 +1,15 @@
 import { Psychart } from 'psychart';
 
-const ps = new Psychart({
+const psychart = new Psychart({
     unitSystem: 'SI',
     dbMax: 50, // [deg C]
     dbMin: 10, // [deg C]
     dpMax: 40, // [deg C]
     flipXY: true,
     yAxis: 'hr', // Humidity Ratio
+    dAxis: 'h', // Enthalpy
     legend: false,
     showAxisNames: true,
 });
 
-window.addEventListener('load', () => {
-    document.body.append(ps.getElement());
-});
+document.body.append(psychart.getElement());